BackCore/requirements.txt
django
djangorestframework==3.15.2
dj-database-url
pika==1.3.2
psycopg2-binary==2.9.9
python-decouple==3.8
retry
